Added a Makefile and added conditionals for the port to the Raspberry Pi
I'm not sure if upstream (or Arduino) even considers adding a Raspberry Pi to its
supported boards so a pull request will probably fail.
Regardless this can be used on a 'Pi right now with an AS5600.

See examples/RaspberryPi/AS5600-capture-images/AS5600-capture-images.cpp
It connects to a ramps 1.4 board running Marlin.
The extruder is used as a rotor with an AS5600 on it.
The 'pi controls the printer with a serial connection via USB.
The idea is to take pictures with the 'Pi camera, rotate a bit. take another picture,
you get the rest.